+++ js/jquery/Component.js
@@ -69,12 +69,12 @@
                 win = $(window),
                 doc = $(document);
 
-            // event window = func      Bind window event
-            // event document = func    Bind document event
-            // ready document = func    Bind DOM ready event
-            // event property = func    Bind event to collection that matches class property
-            // event .class = func      Bind delegated events to class on document
-            // event context .class = func  - Bind delegated events to class within context
+            // event window = func          Bind window event
+            // event document = func        Bind document event
+            // ready document = func        Bind DOM ready event
+            // event property = func        Bind event to collection that matches class property
+            // event .class = func          Bind delegated events to class on document
+            // event context .class = func  Bind delegated events to class within context
             $.each(this.events, function(key, value) {
                 funcs = $.isArray(value) ? value : [value];
                 keys = key.split(' ');
@@ -124,7 +124,7 @@
         },
 
         /**
-         * Disable
+         * Disable the component.
          */
         disable: function() {
             this.enabled = false;
@@ -132,7 +132,7 @@
         },
 
         /**
-         * Enable
+         * Enable the component.
          */
         enable: function() {
             this.enabled = true;