+++ changelog.md
@@ -7,6 +7,7 @@ This update includes documentation for all components.
 
 * Added `isRetina` to the `Toolkit` flags
 * Added `join-classes()` mixin
+* Updated `$.fn.positionTo()` to re-position if the element falls outside the viewport
 * Replaced `currentIndex` with `index` in all components
 * Removed global `previousIndex` properties
 * Removed global `loadingMessage` and `errorMessage` options. Will all use the messages found in `Toolkit`